% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_matrix.R
\name{custom_matrix}
\alias{custom_matrix}
\title{Provide custom data sources for single Band Raster.}
\usage{
custom_matrix(res, src, ...)
}
\arguments{
\item{res}{...}

\item{src}{...}
}
\description{
Working on it...
}
