% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotate.R
\name{fliplit}
\alias{fliplit}
\title{Flip Left-Right}
\usage{
fliplit(x)
}
\arguments{
\item{x}{Matrix}
}
\value{
Flipped matrix
}
\description{
taken from {rayshader}
}
\examples{
#Fake example
}
\keyword{internal}
