% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/texture_to_brick.R
\name{heightmap_to_raster}
\alias{heightmap_to_raster}
\title{Covert a rayshader heightmap to a Raster}
\usage{
heightmap_to_raster(heightmap, out_type = c("raster", "terra", "stars"))
}
\arguments{
\item{texture}{The texture created from a rayshader pipeline.}
}
\value{
An object of class 'RasterBrick' from the {raster} package.
data.
}
\description{
Pretty sure we don't need or want this now...
}
