% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_drape.R
\name{map_drape}
\alias{map_drape}
\title{Get map overlay as array}
\usage{
map_drape(res, src = "wms_virtualearth", alpha = 1, resample = "Average", ...)
}
\arguments{
\item{res}{The desired cell resolution of the matrix in canvas CRS units.}

\item{src}{Default is 'aws'. The server from which to download topographic.}

\item{alpha}{default 1, set transparency of overlay
data.}
}
\description{
retrieve a map matrix for the canvas region. Set the resolution
of the cell size and the elevation source.
}
