% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_drape.R
\name{map_sources}
\alias{map_sources}
\title{View options for Map layers}
\usage{
map_sources()
}
\value{
A character vector with strings that can be used to set the topogrpahy
data source in raytrix::map_drape()
}
\description{
Function to view options that can be used for the \code{src} value of
ratrix::map_drape()
}
