% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canvas.R
\name{raytrix_set_canvas}
\alias{raytrix_set_canvas}
\alias{set_canvas}
\alias{set_canvas_sf}
\alias{set_canvas_centroid}
\alias{get_canvas}
\title{set/get the canvas of the rayshader-ratrix project}
\usage{
set_canvas(.bounds, crs)

set_canvas_sf(.sf, mask = F)

set_canvas_centroid(
  lat,
  long,
  radius = 5000,
  crs = "+proj=longlat +datum=WGS84"
)

get_canvas(res)
}
\arguments{
\item{crs}{...}

\item{.sf}{The sf/sfc object used to set the extent and crs of the canvas}

\item{mask}{Default is F. NOT WORKING YET!}

\item{lat}{...}

\item{long}{...}

\item{radius}{...}

\item{bounds}{vector of length 4. e.g. \code{c(xmin, xmax, ymin, ymax)}}
}
\description{
To enable the downloading of rayshader-readable data for a desired extent,
it is necessary to set the rayshader-ratrix scene canvas using one of the
\code{set_canvas_x()} functions.
}
\details{
To retrieve this information use \code{get_canvas()}
}
