% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotate.R
\name{rotate}
\alias{rotate}
\title{Internal function which is used to rotate matrices and arrays.}
\usage{
rotate(x)
}
\arguments{
\item{x}{Matrix}
}
\value{
Rotated matrix
}
\description{
Internal function which is used to rotate matrices and arrays.
}
\keyword{internal}
