% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtrix_data.R
\name{rtrix_data}
\alias{rtrix_data}
\title{High Level Data download method.}
\usage{
rtrix_data(dsn, res, resample, ..., bands = 1L, dimension)
}
\arguments{
\item{res}{The desired cell resolution of the matrix in canvas CRS units.}

\item{resample}{...}

\item{...}{...}

\item{bands}{...}

\item{dimension}{...}

\item{src}{The server from which to download topographic or map data. See details:}
}
\description{
retrieve a raster using vapour::warp from a desired source.
This function is taken from {gdalio} https://github.com/hypertidy/gdalio/blob/main/R/gdalio.R
}
\details{
See https://github.com/hypertidy/gdalio and https://gdal.org/drivers/raster/wms.html for examples of custom sources from the web. Alternatively, you can download a file and specify the local path.
}
