% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topo_matrix.R
\name{topo_matrix}
\alias{topo_matrix}
\title{Get topography matrix}
\usage{
topo_matrix(res, src = "aws", resample = "CubicSpline", ...)
}
\arguments{
\item{res}{The desired cell resolution of the matrix in canvas CRS units.}

\item{src}{Default is 'aws'. The server from which to download topographic
data.}
}
\description{
retrieve a topographical matrix for the canvas region. Set the resolution
of the cell size and the elevation source.
}
