% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topo_matrix.R
\name{topo_sources}
\alias{topo_sources}
\title{View options for Topography layers}
\usage{
topo_sources()
}
\value{
A character vector with strings that can be used to set the topogrpahy
data source in raytrix::topo_matrix()
}
\description{
Function to view options that can be used for the \code{src} value of
ratrix::topo_matrix()
}
