% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_sf.R
\name{alpha_sf}
\alias{alpha_sf}
\title{create a faded alpha layer for a texture map}
\usage{
alpha_sf(
  texture,
  sf_feature,
  alpha_1_buff,
  kernel_dim = c(9, 9),
  kernel_extent = 9,
  ...
)
}
\arguments{
\item{texture}{...}

\item{sf_feature}{...}

\item{alpha_1_buff}{...}

\item{kernel_dim}{...}

\item{kernel_extent}{...}

\item{...}{passed to \code{rayimage::render_convolution_fft}}
}
\description{
...
}
\details{
...
}
