% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/texture_to_brick.R
\name{heightmap_to_raster}
\alias{heightmap_to_raster}
\title{Covert a rayshader heightmap to a Raster}
\usage{
heightmap_to_raster(heightmap)
}
\arguments{
\item{texture}{The texture created from a rayshader pipeline.}
}
\value{
An object of class 'RasterBrick' from the {raster} package.
data.
}
\description{
retrieve rayshader texture as a RasterBrick. This is handy if you want to
save the texture as to file with spatial reference or if you want  to plot in
an alternative package such as ggplot or tmap etc.
}
