% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_3d_mesh.R
\name{plot_3d_mesh}
\alias{plot_3d_mesh}
\title{plot texture over a 3d mesh}
\usage{
plot_3d_mesh(height_map, texture, zscale = 0.2, lit = F, windowsize = 600, ...)
}
\description{
This is an alternative to rayshader's \code{plot_3d} but rather than plotting the
elevation as a matrix it converts
}
