% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canvas.R
\name{raytrix_set_canvas}
\alias{raytrix_set_canvas}
\alias{set_canvas}
\alias{set_canvas_sf}
\alias{get_canvas}
\title{set/get the canvas of the rayshader-ratrix project}
\usage{
set_canvas(.bounds, crs)

set_canvas_sf(.sf, mask = F)

get_canvas()
}
\arguments{
\item{crs}{The Coordinate reference system of the canvas. Eiher a numeric
EPSG code or a proj string such as "+proj=longlat +datum=WGS84".}

\item{.sf}{The sf/sfc object used to set the extent and crs of the canvas}

\item{mask}{Default is F. NOT WORKING YET!}

\item{bounds}{vector of length 4. e.g. \code{c(xmin, xmax, ymin, ymax)}}
}
\description{
To enable the downloading of rayshader-readable data for a desired extent,
it is necessary to set the rayshader-ratrix scene canvas using one of the
\code{set_canvas_x()} functions.
}
\details{
To retrieve this information use \code{get_canvas()}
}
