% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topo_matrix.R
\name{topo_matrix}
\alias{topo_matrix}
\title{Get topography matrix}
\usage{
topo_matrix(
  res,
  src = "gebco",
  resample = "CubicSpline",
  out_type = "matrix",
  dimension,
  ...
)
}
\arguments{
\item{res}{The desired cell resolution of the matrix in canvas CRS units.}

\item{src}{Default is 'gebco'. The server from which to download topographic
data. Use \code{raytrix::topo_sources()} to view available sources (via {toppgraphy}).
See details.}

\item{resample}{...}

\item{out_type}{...}

\item{dimension}{...}

\item{...}{...}
}
\description{
retrieve a topographical matrix for the canvas region. Set the resolution
of the cell size and the elevation source.
}
\details{
See https://github.com/hypertidy/gdalio and https://gdal.org/drivers/raster/wms.html for examples of custom sources from the web. Alternatively, you can download a file and specify the local path.
}
